cron/pyproject.toml
[project]
name = "cron"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.30.0",
    "oracledb>=3.3.0",
    "polars>=1.31.0",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.42",
]
